% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{assert_IPEDS_fields}
\alias{assert_IPEDS_fields}
\title{Tests whether all the parsed, individual racial data fields are present in
the supplied dataframe/tibble}
\usage{
assert_IPEDS_fields(.data)
}
\arguments{
\item{.data}{A dataframe or tibble}
}
\value{
Boolean: TRUE (if all fields are present), FALSE (if any are missing)
}
\description{
Tests whether all the parsed, individual racial data fields are present in
the supplied dataframe/tibble
}
