% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_old_files.R
\name{delete_old_files}
\alias{delete_old_files}
\title{Keep only the most recent file in a directory that matches a pattern}
\usage{
delete_old_files(path, file_pattern, return_dir = TRUE)
}
\arguments{
\item{path}{Path/directory to be searched for matching files}

\item{file_pattern}{Regular expression to find matching files}

\item{return_dir}{True/False: Should the function return the working directory}
}
\value{
Defaults to the working directory which helps with piped operations
}
\description{
Keep only the most recent file in a directory that matches a pattern
}
