% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize_interval.R
\name{factorize_interval}
\alias{factorize_interval}
\title{Converts a lubridate interval object into a character factor}
\usage{
factorize_interval(x)
}
\arguments{
\item{x}{a lubridate interval object}
}
\description{
Converts a lubridate interval object into a character factor
}
