% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_num_range.R
\name{fct_num_range}
\alias{fct_num_range}
\title{Parse and re-level ambiguous numerical range factors}
\usage{
fct_num_range(x, .range_sep = " - ", .return = "factor")
}
\arguments{
\item{.range_sep}{A string which separates the two halves of the numeric range}

\item{.return}{A string ("factor" or "table") indicating the function output}

\item{.fctr}{A factor or atomic vector that can be coerced into a factor}
}
\value{
Either the input factor re-leveled or the table of levels parsed into columns
}
\description{
By default, \code{fct_num_range()} returns the re-leveled factor
}
\details{
This function takes a factor and typically returns the same but re-leveled so
that ambiguous numeric ranges (e.g. less than 10, 11-15, 16-20, more than 20)
are ordered correctly as factor levels. The function can also be directed
to return a table with the full level parsing, including upper and lower
boundaries. This useful in complex cases where this parser is insufficient.
The function also removes any currency symbols and thousands separators.
It is localized for US - North America where "," is the thousands separator
and "." is the decimal separator.
}
