% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_files_by_date.R
\name{filter_files_by_date}
\alias{filter_files_by_date}
\title{Filter file list based on their date attributes}
\usage{
filter_files_by_date(
  file_names,
  n_return = "all",
  date_filter = "mtime",
  return_what = "paths"
)
}
\arguments{
\item{file_names}{A character vector of fully expanded file paths (not a list)}

\item{n_return}{The number of files to be returned or "all" of them}

\item{date_filter}{Which file date property should be used for filtering?}

\item{return_what}{What should the output format be, a vector of "paths" or a "table" of files and their attributes?}
}
\value{
A list of fully expanded file paths or a table of filtered file names with their attributes
}
\description{
Filter file list based on their date attributes
}
