% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify_app_steps.R
\name{fortify_app_steps}
\alias{fortify_app_steps}
\title{Fortify applicant data with Y/N fields for major application steps}
\usage{
fortify_app_steps(df.omnibus)
}
\arguments{
\item{df.omnibus}{A data.frame of the AMP "Omnibus" report}
}
\value{
The input data.frame with several additional columns for each of the
application steps: complete, interviewed, accepted, matriculated
}
\description{
Fortify applicant data with Y/N fields for major application steps
}
