% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trim_by_n.R
\name{mean_trim_by_n}
\alias{mean_trim_by_n}
\title{Calculate a trimmed mean, removing n values from each end of the sequence}
\usage{
mean_trim_by_n(x, n.trim)
}
\arguments{
\item{x}{A numeric vector}

\item{n.trim}{A number}
}
\value{
The vector mean with 'n.trim' values removed from each end before calculating
}
\description{
Calculate a trimmed mean, removing n values from each end of the sequence
}
