% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_date_on_year.R
\name{norm_date_on_year}
\alias{norm_date_on_year}
\title{Calculates the academic, fiscal, or application year given a date object}
\usage{
norm_date_on_year(
  date,
  acad_year = as.numeric(format(Sys.Date(), "\%Y")),
  anchor_date = "2020-01-01",
  format = "date"
)
}
\arguments{
\item{date}{a date object}

\item{acad_year}{a 4-digit integer giving the academic, fiscal, or admissions cycle year}

\item{anchor_date}{the date on which to normalize the input date(s)}

\item{format}{what the function should return, either a date or count of days from the anchor date}
}
\description{
Calculates the academic, fiscal, or application year given a date object
}
