% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_acad_year.R
\name{parse_acad_year}
\alias{parse_acad_year}
\title{Calculates the academic, fiscal, or application year given a date object}
\usage{
parse_acad_year(date, start = as.Date("1970-07-01"))
}
\arguments{
\item{date}{a date object}

\item{start}{a date object or string coercible into a date}
}
\description{
Calculates the academic, fiscal, or application year given a date object
}
