% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gender.R
\name{parse_gender}
\alias{parse_gender}
\title{Parse gender data and fortify with a standardized, ternary gender field}
\usage{
parse_gender(data, gender.field = gender_identity, parsed.gender = gender)
}
\arguments{
\item{data}{A data.frame or something coercible to this}

\item{gender.field}{An unquoted field within 'data' where gender data lies}

\item{parsed.gender}{An unquoted field name to be appended to 'data'}
}
\value{
The input data.frame fortified with a standardized gender column
}
\description{
Parse gender data and fortify with a standardized, ternary gender field
}
