% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{parse_hispanic_boolean}
\alias{parse_hispanic_boolean}
\title{Parse a binary field of Hispanic/Latino designation}
\usage{
parse_hispanic_boolean(data, hispanic.field)
}
\arguments{
\item{data}{A data.frame}

\item{hispanic.field}{A boolean field}
}
\value{
A data.frame
}
\description{
Parse a binary field of Hispanic/Latino designation
}
