% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{parse_hispanic_code}
\alias{parse_hispanic_code}
\title{Parse the AMCAS Hispanic code field}
\usage{
parse_hispanic_code(data, hispanic.field)
}
\arguments{
\item{data}{A data.frame}

\item{hispanic.field}{Name of Hispanic field from AMCAS}
}
\value{
A data.frame
}
\description{
Parse the AMCAS Hispanic code field
}
