% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{parse_hispanic_text}
\alias{parse_hispanic_text}
\title{Parse a text field containing Hispanic/Latino ethnicity information}
\usage{
parse_hispanic_text(data, hispanic.field)
}
\arguments{
\item{data}{A data.frame}

\item{hispanic.field}{A field name}
}
\value{
A fortified data.frame
}
\description{
Parse a text field containing Hispanic/Latino ethnicity information
}
