% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{parse_ipeds}
\alias{parse_ipeds}
\title{Creates a single, IPEDS-standardized field for race/ethnicity based on the
individual fields created through the 'parse_race' function}
\usage{
parse_ipeds(data)
}
\arguments{
\item{data}{A dataframe or tibble}
}
\value{
Dataframe or tibble fortified with new column
}
\description{
Creates a single, IPEDS-standardized field for race/ethnicity based on the
individual fields created through the 'parse_race' function
}
