% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{parse_race_code}
\alias{parse_race_code}
\title{Parse code-based racial identity field and fortify the dataset with
individual fields for each of the IPEDS racial categories.}
\usage{
parse_race_code(data, race.field, hispanic.field)
}
\arguments{
\item{hispanic.field}{}
}
\value{
Dataframe or tibble fortified with new columns
}
\description{
Parse code-based racial identity field and fortify the dataset with
individual fields for each of the IPEDS racial categories.
}
