% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{parse_urm}
\alias{parse_urm}
\title{Parse the fortified, standardized race and ethnicity columns in the dataset,
and fortify with a new column, calculated URM status.}
\usage{
parse_urm(.data)
}
\arguments{
\item{.data}{A dataframe or tibble}
}
\value{
Dataframe or tibble fortified with new column
}
\description{
Parse the fortified, standardized race and ethnicity columns in the dataset,
and fortify with a new column, calculated URM status.
}
