% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_WI_residency.R
\name{parse_WI_residency}
\alias{parse_WI_residency}
\title{Appends a single Wisconsin residency determination field to the dataset}
\usage{
parse_WI_residency(
  df,
  res_registar = residency_registrar,
  res_amcas = residency_response,
  res_state = residency_state_descr
)
}
\arguments{
\item{df}{A dataframe}

\item{res_registar}{RO residency determination, field in dataframe}

\item{res_amcas}{AMCAS residency response, field inside dataframe}

\item{res_state}{AMCAS state of residence, field inside dataframe}
}
\value{
A dataframe with the Wisconsin residency field appended
}
\description{
Appends a single Wisconsin residency determination field to the dataset
}
