% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_files_by_date.R
\name{pull_files_by_date}
\alias{pull_files_by_date}
\title{Sort selected files in a directory based on their date attribute}
\usage{
pull_files_by_date(
  folder_path = getwd(),
  file_pattern = NULL,
  n_return = NULL,
  date_sort = "mtime",
  skip_temp = TRUE,
  paths_only = TRUE
)
}
\arguments{
\item{folder_path}{A list or character vector of directories to search}

\item{file_pattern}{Regular expression selecting file list to be sorted}

\item{n_return}{How many files should be returned?}

\item{date_sort}{Which file date property should be used for sorting?}

\item{skip_temp}{Should temporary files be skipped? (Almost always "yes")}

\item{paths_only}{Return the full, sorted file paths?}
}
\value{
A list of fully expanded file paths
}
\description{
Sort selected files in a directory based on their date attribute
}
