% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_AMCAS_race_ethn.R
\name{test_race_field}
\alias{test_race_field}
\title{Tests whether the  field is formatted based on the AMCAS / Dept. of Education
racial code standard framework.}
\usage{
test_race_field(.field)
}
\arguments{
\item{.field}{A character vector}
}
\value{
TRUE or FALSE
}
\description{
Tests whether the  field is formatted based on the AMCAS / Dept. of Education
racial code standard framework.
}
