% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_file_repo.R
\name{update_file_repo}
\alias{update_file_repo}
\title{Update archive/repository files with most recent versions}
\usage{
update_file_repo(
  files,
  destination = getwd(),
  action = "copy",
  overwrite = FALSE
)
}
\arguments{
\item{files}{A vector of (full) paths to the files that will be copied/moved}

\item{destination}{An atomic vector giving the destination directory}

\item{action}{File operation (copy or move) to be taken}

\item{overwrite}{Overwrite files with the same name(s) in the destination directory?}
}
\value{
Nothing. Message displayed upon success
}
\description{
Update archive/repository files with most recent versions
}
