% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_UW_colors.R
\name{use_UW_colors}
\alias{use_UW_colors}
\title{Use/Load UW Color Palette}
\usage{
use_UW_colors(palette = "multicolor")
}
\arguments{
\item{palette}{A text string identifying the palette to loaded, one of the following: multicolor, compare1/2/2uw, pairwise}
}
\value{
A string vector of hexidecimal colors
}
\description{
Use/Load UW Color Palette
}
