% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_files_by_date.R
\name{sort_files_by_date}
\alias{sort_files_by_date}
\title{Sort file list based on their date attributes}
\usage{
sort_files_by_date(
  file_names,
  n_return = "all",
  date_sort = "mtime",
  return_what = "paths"
)
}
\arguments{
\item{file_names}{A character vector of fully expanded file paths (not a list)}

\item{n_return}{The number of files to be returned or "all" of them}

\item{date_sort}{Which file date property should be used for sorting?}

\item{return_what}{What should the output format be, a vector of "paths" or a "table" of files and their attributes?}
}
\value{
A list of fully expanded file paths or a table of sorts files names with their attributes
}
\description{
Sort file list based on their date attributes
}
