% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_files_by_date.R
\name{sort_files_by_date}
\alias{sort_files_by_date}
\title{Title}
\usage{
sort_files_by_date(
  folder_path = getwd(),
  file_pattern = NULL,
  n_return = NULL,
  date_sort = "mtime",
  skip_temp = TRUE,
  paths_only = TRUE
)
}
\arguments{
\item{folder_path}{A directory path}

\item{file_pattern}{Regular expression selecting file list to be sorted}

\item{n_return}{How many files should be returned?}

\item{date_sort}{Which file date property should be used for sorting?}

\item{skip_temp}{Should temporary files be skipped? (Almost always "yes")}

\item{paths_only}{Return the full, sorted file paths?}
}
\value{
A list of fully expanded file paths
}
\description{
Title
}
